rename operation.rb to match class name
