+++ lib/zakuro/result/operation/bundle.rb
