+++ lib/zakuro/result/result.rb
@@ -4,7 +4,7 @@ require 'json'
 
 require_relative './data/single_day'
 
-require_relative './operation/operation'
+require_relative './operation/bundle'
 
 require_relative './core'
